05_feature_engineering/reddit_scheduler_5_get_rep_submissions/requirements.txt
functions-framework==3.*

pandas
fsspec
google-cloud-storage
gcsfs

pyarrow
fastparquet

numpy